==> code/llm/3-llm-tag-validating.ipynb <==
import glob
import pandas as pd
import re

from langchain_core.prompts import ChatPromptTemplate
from langchain_core.pydantic_v1 import BaseModel, Field
from langchain_openai import ChatOpenAI

OPENAI_API_KEY = "YOUR API KEY GOES HERE"
OPENAI_API_MODEL = 'gpt-3.5-turbo-0125'

# Get the classified files
files = glob.glob("../../output/classified-llm/*.csv")

df = pd.concat([pd.read_csv(file) for file in files]).reset_index(drop=True)

# Removes columns that we don't need
# df = df.drop(columns=['Unnamed: 0.1', 'Unnamed: 0', 'level_0', 'index'])

df['level_0'].value_counts()

# Let's take a look at all the retrieved tags

# All tags
all_tags = df.topic.str.split(',') # Split in the commas
flattened_list = [item.strip() for sublist in all_tags for item in sublist] # Makes one big list
all_tags_df = pd.DataFrame(flattened_list).rename(columns={0:'tag'}) # Dataframe with one row for wich time a tag appeared

display(all_tags_df)

# Total tags applied
all_tags_df.tag.shape[0]

# Total tags applied
all_tags_df.tag.unique().shape[0]

# That seems like a disproportionate ammount of tags.
# Some of them might be applying to one single speech.
# Let's take a look at the value counts.
display(all_tags_df.tag.value_counts(), all_tags_df.tag.value_counts(normalize=True))

# The top entries concentrate most of the speeches.
# Let's see exactly how many.

# Counts the tags
value_counts = all_tags_df.tag.value_counts()

# Get the 100 most common ones
frequent_tags = value_counts.head(100).index

# Taking a look at the substrings. Let's remove the mentions to specific
# places and stick with general themes and issues.
frequent_tags

frequent_tags = frequent_tags.drop(['European Union', 'Russia', 'Ukraine', 'China', 'Europe', 'EU'])

frequent_tags

# On the original dataframe, select all the speeches that have at least one of this tags
pattern = "|".join(frequent_tags)

# Use str.contains to check if any of the strings are present
df['contains_string'] = df['topic'].str.contains(pattern, case=False, regex=True)

# How many speeches have at least one tag that is among the most common?
df['contains_string'].value_counts(normalize=True)

# 98.5% of the speeches contain at least one this tags!!!
# Let's look at the ones that doesn't.

# Overall, the share of English speeches in the group 
# for which we DON'T have data is nearly 10% LARGER to the one we have
# data for. This could show some bias: the classifier has
# more problems dealing with foreign languages.

df[df.contains_string].language.str.lower().value_counts(normalize=True)

df[~df.contains_string].language.str.lower().value_counts(normalize=True)

# Share of speeches by language that didn't get any valid tags
(df[~df.contains_string].language.str.lower().value_counts() / \
df.language.str.lower().value_counts()) * 100

# No language was hugely affected

# Removes entries with invalid strings
df = df[df.contains_string]

# Let's create a regex pattern with capture groups to get the frequent tags that were applied into each.
pattern = '|'.join(f'(?:{substring})' for substring in frequent_tags)

def find_substrings(text, pattern):
    matches = re.findall(pattern, text, re.IGNORECASE)
    return ', '.join(matches)

# Apply the function to the DataFrame
df['valid_tags'] = df['topic'].apply(lambda x: find_substrings(x, pattern))

# Clean's data up for further processing
df = df[['level_0', 'speech', 'date', 'term', 'year', 'speech_length_in_characters', 'language', 'valid_tags']]

df = df.reset_index(drop=True)

# Saves
df.to_feather("../../output/llm-validated/validated-data.feather")

frequent_tags
